+++ src/handlers/scheduler_handler.go
@@ -10,8 +10,8 @@ import (
 	lgr "github.com/SimplyVC/oasis_api_server/src/logger"
 	"github.com/SimplyVC/oasis_api_server/src/responses"
 	"github.com/SimplyVC/oasis_api_server/src/rpc"
-	common_namespace "github.com/oasislabs/oasis-core/go/common"
-	scheduler "github.com/oasislabs/oasis-core/go/scheduler/api"
+	common_namespace "github.com/oasisprotocol/oasis-core/go/common"
+	scheduler "github.com/oasisprotocol/oasis-core/go/scheduler/api"
 )
 
 // loadSchedulerClient loads scheduler client and returns it
@@ -52,7 +52,7 @@ func GetValidators(w http.ResponseWriter, r *http.Request) {
 
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
-			Error: "Unexepcted value found, height needs to be " +
+			Error: "Unexpected value found, height needs to be " +
 				"string of int!"})
 		return
 	}
@@ -114,7 +114,7 @@ func GetCommittees(w http.ResponseWriter, r *http.Request) {
 	if height == -1 {
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
-			Error: "Unexepcted value found, height needs to be " +
+			Error: "Unexpected value found, height needs to be " +
 				"string of int!"})
 		return
 	}
@@ -205,7 +205,7 @@ func GetSchedulerStateToGenesis(w http.ResponseWriter, r *http.Request) {
 
 		// Stop code here no need to establish connection and reply
 		json.NewEncoder(w).Encode(responses.ErrorResponse{
-			Error: "Unexepcted value found, height needs to be " +
+			Error: "Unexpected value found, height needs to be " +
 				"string of int!"})
 		return
 	}