+++ src/handlers/sentry_handler.go
@@ -10,7 +10,7 @@ import (
 	lgr "github.com/SimplyVC/oasis_api_server/src/logger"
 	"github.com/SimplyVC/oasis_api_server/src/responses"
 	"github.com/SimplyVC/oasis_api_server/src/rpc"
-	sentry "github.com/oasislabs/oasis-core/go/sentry/api"
+	sentry "github.com/oasisprotocol/oasis-core/go/sentry/api"
 )
 
 // loadSentryClient loads sentry client and returns it